ArmorIQ/armorclaw-shield-2026/requirements.txt
google-generativeai>=0.5.0
alpaca-trade-api>=3.0.0
firecrawl-py>=0.0.16
mcp>=1.0.0
pyyaml>=6.0
fastapi>=0.110.0
uvicorn>=0.27.0
rich>=13.7.0
python-dotenv>=1.0.0
pytz>=2024.1
requests>=2.31.0
websockets>=12.0